move footer/header
