+++ src/routes/+layout.svelte
@@ -1,7 +1,7 @@
 <script>
 	import '../app.css';
-	import Header from '$lib/Header.svelte';
-	import Footer from '$lib/Footer.svelte';
+	import Header from '$lib/components/Header.svelte';
+	import Footer from '$lib/components/Footer.svelte';
 	import { fly } from 'svelte/transition';
 	import { nav_expanded, blur } from '$lib/app';
